feat : 예외 처리 추가 추가로 가독성을 위한 리팩토링 같이 진행

+++ src/main/java/gift/Service/WishService.java
@@ -1,6 +1,7 @@
 package gift.Service;
 
 import gift.Exception.AuthorizedException;
+import gift.Exception.ProductNotFoundException;
 import gift.Model.*;
 import gift.Repository.ProductRepository;
 import gift.Repository.MemberRepository;
@@ -15,22 +16,32 @@ import java.util.Optional;
 
 @Service
 public class WishService {
-    @Autowired
-    private WishRepository wishRepository;
-    @Autowired
-    private MemberRepository memberRepository;
-    @Autowired
-    private ProductRepository productRepository;
-    @Autowired
-    private JwtTokenProvider jwtTokenProvider;
+    private final WishRepository wishRepository;
+    private final MemberRepository memberRepository;
+    private final ProductRepository productRepository;
+    private final JwtTokenProvider jwtTokenProvider;
+
+    public WishService(WishRepository wishRepository, MemberRepository memberRepository, ProductRepository productRepository, JwtTokenProvider jwtTokenProvider){
+        this.wishRepository = wishRepository;
+        this.memberRepository = memberRepository;
+        this.productRepository = productRepository;
+        this.jwtTokenProvider = jwtTokenProvider;
+    }
 
     public void add(String token, String name){
         String email = jwtTokenProvider.getEmailFromToken(token);
         Optional<Member> memberOptional = memberRepository.findByEmail(email);
         Optional<Product> productOptional = productRepository.findByName(name);
-        if(memberOptional.isEmpty()) throw new AuthorizedException();
-        Product product = productOptional.get();
+        if(memberOptional.isEmpty()) {
+            throw new AuthorizedException();
+        }
+
+        if(productOptional.isEmpty()){
+            throw new ProductNotFoundException();
+        }
+
         Member member = memberOptional.get();
+        Product product = productOptional.get();
         if(!member.getRole().equals(Role.ADMIN) && !member.getRole().equals(Role.CONSUMER))
             throw new AuthorizedException();
 
@@ -41,12 +52,20 @@ public class WishService {
         String email = jwtTokenProvider.getEmailFromToken(token);
         Optional<Member> memberOptional = memberRepository.findByEmail(email);
         Optional<Product> productOptional = productRepository.findByName(name);
+
+        if(memberOptional.isEmpty()) {
+            throw new AuthorizedException();
+        }
+
+        if(productOptional.isEmpty()){
+            throw new ProductNotFoundException();
+        }
+
         Product product = productOptional.get();
-        if(memberOptional.isEmpty())
-            throw new AuthorizedException();
         Member member = memberOptional.get();
-        if(!member.getRole().equals(Role.ADMIN) && !member.getRole().equals(Role.CONSUMER))
+        if(!member.getRole().equals(Role.ADMIN) && !member.getRole().equals(Role.CONSUMER)) {
             throw new AuthorizedException();
+        }
 
         wishRepository.delete(wishRepository.findByMemberIdAndProductId(member.getId(), product.getId()));
     }
@@ -55,18 +74,22 @@ public class WishService {
         String email = jwtTokenProvider.getEmailFromToken(token);
         Optional<Member> memberOptional = memberRepository.findByEmail(email);
 
-        if(memberOptional.isEmpty())
+        if(memberOptional.isEmpty()) {
             throw new AuthorizedException();
+        }
         Member member = memberOptional.get();
 
-        if(!member.getRole().equals(Role.ADMIN) && !member.getRole().equals(Role.CONSUMER))
+        if(!member.getRole().equals(Role.ADMIN) && !member.getRole().equals(Role.CONSUMER)) {
             throw new AuthorizedException();
+        }
 
         List<Wish> wishes = wishRepository.findByMemberId(member.getId());
         List<String> productNames = new ArrayList<>();
+
         for(Wish w : wishes){
             productNames.add(productRepository.findById(w.getProductId()).get().getName());
         }
+
         return productNames;
     }
 }
